+++ spring-web-plus-tool/src/main/java/com/github/pdaodao/springwebplus/tool/db/dialect/pg/PgDialect.java
@@ -55,7 +55,7 @@ public class PgDialect extends BaseDbDialect {
     }
 
     @Override
-    public String pageSql(String sql, Integer offset, Integer size) {
+    public String pageSql(String sql, Long offset, Long size) {
         if (offset == null || offset < 1) {
             return DbUtil.pageSqlWrap(sql, "limit " + size);
         }
